+++ components/scg/src/gwf-object-info-reader.js
@@ -10,7 +10,7 @@ GwfObjectInfoReader = {
         "node/const/perm/tuple": sc_type_const | sc_type_node | sc_type_node_tuple,
         "node/const/perm/struct": sc_type_const | sc_type_node | sc_type_node_structure,
         "node/const/perm/role": sc_type_const | sc_type_node | sc_type_node_role,
-        "node/const/perm/relation": sc_type_const | sc_type_node | sc_type_node_norole,
+        "node/const/perm/relation": sc_type_const | sc_type_node | sc_type_node_no_role,
         "node/const/perm/terminal": sc_type_const | sc_type_node | sc_type_node_material,
         "node/const/perm/group": sc_type_const | sc_type_node | sc_type_node_class,
 
@@ -18,15 +18,15 @@ GwfObjectInfoReader = {
         "node/var/perm/tuple": sc_type_var | sc_type_node | sc_type_node_tuple,
         "node/var/perm/struct": sc_type_var | sc_type_node | sc_type_node_structure,
         "node/var/perm/role": sc_type_var | sc_type_node | sc_type_node_role,
-        "node/var/perm/relation": sc_type_var | sc_type_node | sc_type_node_norole,
+        "node/var/perm/relation": sc_type_var | sc_type_node | sc_type_node_no_role,
         "node/var/perm/terminal": sc_type_var | sc_type_node | sc_type_node_material,
         "node/var/perm/group": sc_type_var | sc_type_node | sc_type_node_class,
 
         "node/const/general_node": sc_type_node_link | sc_type_const,
         "node/var/general_node": sc_type_node_link | sc_type_var,
 
-        "pair/const/fuz/temp/orient/membership": sc_type_const | sc_type_membership_arc | sc_type_fuz_arc | sc_type_temp_arc,
-        "pair/const/fuz/perm/orient/membership": sc_type_const | sc_type_membership_arc | sc_type_fuz_arc | sc_type_perm_arc,
+        "pair/const/fuz/temp/orient/membership": sc_type_const | sc_type_membership_arc | sc_type_fuz_arc,
+        "pair/const/fuz/perm/orient/membership": sc_type_const | sc_type_membership_arc | sc_type_fuz_arc,
         "pair/const/pos/temp/orient/membership": sc_type_const | sc_type_membership_arc | sc_type_pos_arc | sc_type_temp_arc,
         "pair/const/pos/perm/orient/membership": sc_type_const | sc_type_membership_arc | sc_type_pos_arc | sc_type_perm_arc,
         "pair/const/neg/temp/orient/membership": sc_type_const | sc_type_membership_arc | sc_type_neg_arc | sc_type_temp_arc,
@@ -36,9 +36,12 @@ GwfObjectInfoReader = {
         "pair/-/-/-/orient": sc_type_common_arc,
 
         "pair/var/-/perm/orient": sc_type_common_arc | sc_type_var,
-        "pair/var/fuz/perm/orient/membership": sc_type_var | sc_type_membership_arc | sc_type_fuz_arc | sc_type_perm_arc,
+        "pair/var/fuz/perm/orient/membership": sc_type_var | sc_type_membership_arc | sc_type_fuz_arc,
+        "pair/var/pos/temp/orient/membership": sc_type_var | sc_type_membership_arc | sc_type_pos_arc | sc_type_temp_arc,
         "pair/var/pos/perm/orient/membership": sc_type_var | sc_type_membership_arc | sc_type_pos_arc | sc_type_perm_arc,
+        "pair/var/neg/temp/orient/membership": sc_type_var | sc_type_membership_arc | sc_type_neg_arc | sc_type_temp_arc,
         "pair/var/neg/perm/orient/membership": sc_type_var | sc_type_membership_arc | sc_type_neg_arc | sc_type_perm_arc,
+        "pair/var/-/perm/orient": sc_type_common_arc | sc_type_var,
         "pair/var/-/perm/noorien": sc_type_common_edge | sc_type_var,
         "pair/-/-/-/noorient": sc_type_common_edge
     },